App Skeleton
Added dependency (electron.js) and related script (start).
Added index.html, containing the HTML of the main window of the app,
and main.js, interracting with electron to create the app in a chromium
window.

The window has no Chromium frame, which makes it render without any
titlebar or border.